[T-2] INR and GBP against EUR from 01-01-19 to 31-01-19  Multi-line graph without any external library
